Set response_mode to form_post
React to changes Apple have made to the sign-in service, and use form_post as the response mode.
This requires a dirty hack in the short term to copy the form parameters into the query string and swap it out so that the base class' implementation to handle things can be used as-is.